Added a SQLite database to keep track of leaderboard, need to change the insert to upsert

+++ src/main/scala/helper/Comparer.scala
@@ -39,10 +39,10 @@ object Comparer {
   }
 
   // Returns the similarity between two files or empty when submission was empty
-  def similarity(fileModel: File, fileToCompare: File): Option[BigDecimal] = {
+  def similarity(fileModel: File, fileToCompare: File): Option[Double] = {
 
       compareFiles(fileModel, fileToCompare) match {
-        case Some(similarity) => Some(BigDecimal(similarity*100).setScale(3, BigDecimal.RoundingMode.HALF_UP))
+        case Some(similarity) => Some(BigDecimal(similarity*100).setScale(3, BigDecimal.RoundingMode.HALF_UP).toDouble)
         case None => None
       }
